backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
eventlet==0.35.1
PyJWT==2.8.0
werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
